removed singleton and API
