+++ meter/bosch_bpts5_hybrid.go
@@ -14,12 +14,17 @@ package meter
 
 import (
 	"errors"
+	"fmt"
+	"net/http"
+	"net/http/cookiejar"
+	"strconv"
 	"strings"
 	"time"
 
 	"github.com/evcc-io/evcc/api"
-	"github.com/evcc-io/evcc/meter/bosch"
 	"github.com/evcc-io/evcc/util"
+	"github.com/evcc-io/evcc/util/request"
+	"github.com/evcc-io/evcc/util/transport"
 )
 
 // Example config:
@@ -38,8 +43,19 @@ import (
 //   usage: battery
 
 type BoschBpts5Hybrid struct {
-	api   *bosch.API
-	usage string
+	*request.Helper
+	usage                 string
+	wuSid                 string
+	uri                   string
+	updated               time.Time
+	cache                 time.Duration
+	logger                *util.Logger
+	CurrentBatterySoc     float64
+	SellToGrid            float64
+	BuyFromGrid           float64
+	PvPower               float64
+	BatteryChargePower    float64
+	BatteryDischargePower float64
 }
 
 func init() {
@@ -71,16 +87,21 @@ func NewBoschBpts5HybridFromConfig(other map[string]interface{}) (api.Meter, err
 func NewBoschBpts5Hybrid(uri, usage string, cache time.Duration) (api.Meter, error) {
 	log := util.NewLogger("bosch-bpts5-hybrid")
 
-	instance, exists := bosch.Instances.LoadOrStore(uri, bosch.NewLocal(log, uri, cache))
-	if !exists {
-		if err := instance.(*bosch.API).Login(); err != nil {
-			return nil, err
-		}
+	m := &BoschBpts5Hybrid{
+		Helper: request.NewHelper(log),
+		usage:  strings.ToLower(usage),
+		uri:    util.DefaultScheme(strings.TrimSuffix(uri, "/"), "http"),
+		cache:  cache,
+		logger: log,
 	}
 
-	m := &BoschBpts5Hybrid{
-		api:   instance.(*bosch.API),
-		usage: strings.ToLower(usage),
+	// ignore the self signed certificate
+	m.Client.Transport = request.NewTripper(log, transport.Insecure())
+	// create cookie jar to save login tokens
+	m.Client.Jar, _ = cookiejar.New(nil)
+
+	if err := m.login(); err != nil {
+		return nil, err
 	}
 
 	// decorate api.BatterySoC
@@ -92,17 +113,31 @@ func NewBoschBpts5Hybrid(uri, usage string, cache time.Duration) (api.Meter, err
 	return decorateBoschBpts5Hybrid(m, batterySoC), nil
 }
 
+func (c *BoschBpts5Hybrid) login() (err error) {
+	req, err := request.New(http.MethodGet, c.uri, nil, nil)
+	if err != nil {
+		return err
+	}
+
+	body, err := c.DoBody(req)
+	if err != nil {
+		return err
+	}
+
+	return extractWuiSidFromBody(c, string(body))
+}
+
 // CurrentPower implements the api.Meter interface
 func (m *BoschBpts5Hybrid) CurrentPower() (float64, error) {
-	status, err := m.api.Status()
+	err := m.status()
 
 	switch m.usage {
 	case "grid":
-		return status.BuyFromGrid - status.SellToGrid, err
+		return m.BuyFromGrid - m.SellToGrid, err
 	case "pv":
-		return status.PvPower, err
+		return m.PvPower, err
 	case "battery":
-		return status.BatteryDischargePower - status.BatteryChargePower, err
+		return m.BatteryDischargePower - m.BatteryChargePower, err
 	default:
 		return 0, err
 	}
@@ -110,6 +145,117 @@ func (m *BoschBpts5Hybrid) CurrentPower() (float64, error) {
 
 // batterySoC implements the api.Battery interface
 func (m *BoschBpts5Hybrid) batterySoC() (float64, error) {
-	status, err := m.api.Status()
-	return status.CurrentBatterySoc, err
+	err := m.status()
+	return m.CurrentBatterySoc, err
+}
+
+//////////// value retrieval ////////////////
+
+func (c *BoschBpts5Hybrid) status() error {
+	var err error
+	if time.Since(c.updated) > c.cache {
+		if err = c.updateValues(); err == nil {
+			c.updated = time.Now()
+		}
+	}
+	return err
+}
+
+//////////// helpers ////////////////
+
+func extractWuiSidFromBody(c *BoschBpts5Hybrid, body string) error {
+	index := strings.Index(body, "WUI_SID=")
+
+	if index < 0 {
+		c.wuSid = ""
+		return fmt.Errorf("error while extracting wui sid. body was= %s", body)
+	}
+
+	c.wuSid = body[index+9 : index+9+15]
+
+	c.logger.DEBUG.Println("extractWuiSidFromBody: result=", c.wuSid)
+
+	return nil
+}
+
+func (c *BoschBpts5Hybrid) updateValues() error {
+	data := "action=get.hyb.overview&flow=1"
+
+	uri := c.uri + "/cgi-bin/ipcclient.fcgi?" + c.wuSid
+	req, err := request.New(http.MethodPost, uri, strings.NewReader(data), map[string]string{
+		"Content-Type": "text/plain",
+	})
+
+	if err != nil {
+		return err
+	}
+
+	body, err := c.DoBody(req)
+	if err != nil {
+		return err
+	}
+
+	return extractValues(c, string(body))
+}
+
+func parseWattValue(inputString string) (float64, error) {
+	if len(strings.TrimSpace(inputString)) == 0 || strings.Contains(inputString, "nbsp;") {
+		return 0.0, nil
+	}
+
+	numberString := strings.TrimSpace(strings.ReplaceAll(strings.ReplaceAll(inputString, "kW", " "), "von", " "))
+
+	res, err := strconv.ParseFloat(numberString, 64)
+
+	return res * 1000.0, err
+}
+
+func extractValues(c *BoschBpts5Hybrid, body string) error {
+	if strings.Contains(body, "session invalid") {
+		c.logger.DEBUG.Println("extractValues: Session invalid. Performing Re-login")
+		return c.login()
+	}
+
+	values := strings.Split(body, "|")
+
+	if len(values) < 14 {
+		return fmt.Errorf("extractValues: response has not enough values")
+	}
+
+	soc, err := strconv.Atoi(values[3])
+
+	if err != nil {
+		return fmt.Errorf("extractValues: error during value parsing 1: %s", err)
+	}
+
+	c.CurrentBatterySoc = float64(soc)
+	c.SellToGrid, err = parseWattValue(values[11])
+
+	if err != nil {
+		return fmt.Errorf("extractValues: error during value parsing 2: %s", err)
+	}
+
+	c.BuyFromGrid, err = parseWattValue(values[14])
+
+	if err != nil {
+		return fmt.Errorf("extractValues: error during value parsing 3: %s", err)
+	}
+
+	c.PvPower, err = parseWattValue(values[2])
+
+	if err != nil {
+		return fmt.Errorf("extractValues: error during value parsing 4: %s", err)
+	}
+
+	c.BatteryChargePower, err = parseWattValue(values[10])
+
+	if err != nil {
+		return fmt.Errorf("extractValues: error during value parsing 5: %s", err)
+	}
+
+	c.BatteryDischargePower, err = parseWattValue(values[13])
+
+	c.logger.DEBUG.Println("extractValues: batterieLadeStrom=", c.BatteryChargePower, ";currentBatterySocValue=", c.CurrentBatterySoc, ";einspeisung=", c.SellToGrid, ";pvLeistungWatt=", c.PvPower, ";strombezugAusNetz=", c.BuyFromGrid, ";verbrauchVonBatterie=", c.BatteryDischargePower)
+
+	return err
 }
